requirements.txt
pandas
duckdb
pyarrow
apache-airflow

